Bug fix: Menu is available when switching between plan and work views
Added a view observer which detects when the view has changed, and builds the menu in the appropriate place in the DOM. There are two sets of Quick Filters(!), one for the work view and one for the plan view, and we now ensure that the menu is available in both views or after switching between them.

Also added code to blur focus from the menu after choosing a filter so that JIRA's hotkeys continue to work instead of being sent to the menu.

Added non-working code to update the menu state after changing views to show the currently active assignee filter. Will fix this later. (See TODO comments.)